requirements.txt
streamlit==1.32.0
openai==1.12.0
Pillow==10.2.0
requests==2.31.0
httpx==0.27.0 